requirements.txt
streamlit==1.57.0
psycopg2-binary==2.9.12
boto3>=1.34.0
pandas>=2.0.0
